+++ projects/ngx-simple-gallery/src/lib/core/constants.ts
@@ -1,6 +1,6 @@
 export const Constants = {
   defaultEmptyMessage: 'Empty gallery, no images provided.',
-  defaultShowThumbnailList: true,
-  defaultThumbnailSize: 160,
-  defaultThumbnailStartIndex: 0,
+  defaultGalleryThumbnailSize: 160,
+  defaultModalStartIndex: 0,
+  defaultShowModalThumbnailList: true,
 } as const;