+++ projects/ngx-simple-gallery/src/lib/core/model/lib-config.ts
@@ -1,0 +1,6 @@
+export interface LibConfig {
+  emptyMessage: string;
+  galleryThumbnailSize: number;
+  modalStartIndex: number;
+  showModalThumbnailList: boolean;
+}