+++ projects/ngx-simple-gallery/src/lib/core/model/modal-config.ts
@@ -1,4 +1,0 @@
-export interface ModalConfig {
-  showModalThumbnailList?: boolean;
-  startIndex?: number;
-}