+++ projects/ngx-simple-gallery/src/lib/core/service/gallery.service.ts
@@ -1,23 +1,24 @@
 import { computed, Injectable, signal } from '@angular/core';
 import { GalleryItem } from '../model/gallery-item';
 import { Constants } from '../constants';
-import { ModalConfig } from '../model/modal-config';
+import { GalleryConfig } from '../model/gallery-config';
 import { ConfigUtils } from '../utils/config-utils';
+import { LibConfig } from '../model/lib-config';
 
 @Injectable({
   providedIn: 'root',
 })
 export class GalleryService {
+  private readonly libConfig = signal<LibConfig>(ConfigUtils.defaultLibConfig());
   private readonly isLoading = signal<boolean>(false);
-  private readonly itemNdx = signal<number>(Constants.defaultThumbnailStartIndex);
+  private readonly itemNdx = signal<number>(Constants.defaultModalStartIndex);
   private readonly items = signal<GalleryItem[]>([]);
-  private readonly showThumbnailList = signal<boolean>(Constants.defaultShowThumbnailList);
 
   public readonly imageSource = computed(() => this.items().at(this.itemNdx())?.src ?? '');
   public readonly galleryItems = this.items.asReadonly();
   public readonly getIsLoading = this.isLoading.asReadonly();
   public readonly getItemIndex = this.itemNdx.asReadonly();
-  public readonly getShowThumbnailList = this.showThumbnailList.asReadonly();
+  public readonly getLibConfig = this.libConfig.asReadonly();
 
   public setGalleryItems(items: GalleryItem[]): void {
     this.items.set(items);
@@ -28,12 +29,22 @@ export class GalleryService {
     this.itemNdx.set(ConfigUtils.normalizeStartIndex(index, this.galleryItems().length));
   }
 
-  public applyModalConfig(modalConfig: ModalConfig): void {
-    if (modalConfig.showModalThumbnailList !== undefined) {
-      this.showThumbnailList.set(modalConfig.showModalThumbnailList);
-    }
-    if (modalConfig.startIndex !== undefined) {
-      this.setItemIndex(modalConfig.startIndex);
+  public applyGalleryConfig(galleryConfig: GalleryConfig): void {
+    this.libConfig.set({
+      emptyMessage: galleryConfig.emptyMessage !== undefined ? galleryConfig.emptyMessage : Constants.defaultEmptyMessage,
+      galleryThumbnailSize:
+        galleryConfig.galleryThumbnailSize !== undefined
+          ? galleryConfig.galleryThumbnailSize
+          : Constants.defaultGalleryThumbnailSize,
+      modalStartIndex:
+        galleryConfig.modalStartIndex !== undefined ? galleryConfig.modalStartIndex : Constants.defaultModalStartIndex,
+      showModalThumbnailList:
+        galleryConfig.showModalThumbnailList !== undefined
+          ? galleryConfig.showModalThumbnailList
+          : Constants.defaultShowModalThumbnailList,
+    });
+    if (galleryConfig.modalStartIndex !== undefined) {
+      this.setItemIndex(galleryConfig.modalStartIndex);
     }
   }