+++ projects/ngx-simple-gallery/src/lib/open-gallery/open-gallery.directive.ts
@@ -2,7 +2,7 @@ import { Directive, ElementRef, HostListener, inject, Input, input, Renderer2 } 
 import { ShowcaseDialogComponent } from '../showcase-dialog/showcase-dialog.component';
 import { GalleryService } from '../core/service/gallery.service';
 import { Dialog } from '@angular/cdk/dialog';
-import { ModalConfig } from '../core/model/modal-config';
+import { GalleryConfig } from '../core/model/gallery-config';
 import { GalleryItem } from '../core/model/gallery-item';
 
 @Directive({
@@ -23,8 +23,8 @@ export class OpenGalleryDirective {
   }
 
   @Input()
-  set modalConfig(modalConfig: ModalConfig) {
-    this.galleryService.applyModalConfig(modalConfig);
+  set galleryConfig(galleryConfig: GalleryConfig) {
+    this.galleryService.applyGalleryConfig(galleryConfig);
   }
 
   @HostListener('click') onClick(): void {
@@ -35,7 +35,7 @@ export class OpenGalleryDirective {
       this.elementRef.nativeElement.style.backgroundColor = 'red';
       this.elementRef.nativeElement.style.padding = '0.5rem 1rem 0.5rem 1rem';
       this.elementRef.nativeElement.style.fontSize = '1.5rem';
-      this.renderer.setProperty(this.elementRef.nativeElement, 'innerHTML', '<p>Gallery is empty. Provide items through [openGallery]="arrayOfGalleryItems"</p>');
+      this.renderer.setProperty(this.elementRef.nativeElement, 'innerHTML', `<p>${this.galleryService.getLibConfig().emptyMessage}</p>`);
     }
   }
 }