+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { GalleryItem, ModalConfig, NgxSimpleGalleryComponent, OpenGalleryDirective } from 'ngx-simple-gallery';
+import { GalleryItem, GalleryConfig, NgxSimpleGalleryComponent, OpenGalleryDirective } from 'ngx-simple-gallery';
 
 @Component({
   selector: 'app-root',
@@ -9,7 +9,7 @@ import { GalleryItem, ModalConfig, NgxSimpleGalleryComponent, OpenGalleryDirecti
   styleUrl: './app.component.css',
 })
 export class AppComponent {
-  public modalConfig: ModalConfig = { showModalThumbnailList: false, startIndex: 2 };
+  public galleryConfig: GalleryConfig = { emptyMessage: 'Images not found', galleryThumbnailSize: 140, showModalThumbnailList: true, modalStartIndex: 2 };
 
   public galleryItems: GalleryItem[] = [
     {